refactor(user_center): 移除废弃的数据库脚本及多余导包，优化代码可读性

+++ core/src/main/java/io/github/lionheartlattice/entity/user_center/po/Dept.java
@@ -9,6 +9,7 @@ import io.github.lionheartlattice.entity.parent.ParentClientEntity;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 /**
@@ -30,7 +31,7 @@ public class Dept extends ParentClientEntity<Dept, DeptProxy> implements ProxyEn
      */
     @Schema(description = "部门ID", requiredMode = Schema.RequiredMode.REQUIRED)
     @Column(primaryKey = true, value = "id", primaryKeyGenerator = SnowflakePrimaryKeyGenerator.class)
-    private Long id;
+    private BigDecimal id;
 
     /**
      * 部门名称
@@ -42,13 +43,13 @@ public class Dept extends ParentClientEntity<Dept, DeptProxy> implements ProxyEn
      * 父级ID
      */
     @Schema(description = "父级ID", requiredMode = Schema.RequiredMode.REQUIRED)
-    private Long pid;
+    private BigDecimal pid;
 
     /**
      * 默认分配角色ID
      */
     @Schema(description = "默认分配角色ID")
-    private Long defaultRoleId;
+    private BigDecimal defaultRoleId;
 
     /**
      * 层级
@@ -91,13 +92,13 @@ public class Dept extends ParentClientEntity<Dept, DeptProxy> implements ProxyEn
      * 创建人ID
      */
     @Schema(description = "创建人ID")
-    private Long createId;
+    private BigDecimal createId;
 
     /**
      * 更新人ID
      */
     @Schema(description = "更新人ID")
-    private Long updateId;
+    private BigDecimal updateId;
 
     /**
      * 更新时间
@@ -109,6 +110,6 @@ public class Dept extends ParentClientEntity<Dept, DeptProxy> implements ProxyEn
      * 租户ID
      */
     @Schema(description = "租户ID")
-    private Long tenantId;
+    private BigDecimal tenantId;
 
 }
