+++ core/src/main/java/io/github/lionheartlattice/entity/user_center/po/Menu.java
@@ -10,6 +10,7 @@ import io.github.lionheartlattice.entity.parent.ParentClientEntity;
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -32,13 +33,13 @@ public class Menu extends ParentClientEntity<Menu, MenuProxy> implements ProxyEn
      */
     @Schema(description = "菜单表id", requiredMode = Schema.RequiredMode.REQUIRED)
     @Column(primaryKey = true, value = "id", primaryKeyGenerator = SnowflakePrimaryKeyGenerator.class)
-    private Long id;
+    private BigDecimal id;
 
     /**
      * 父级id
      */
     @Schema(description = "父级id", requiredMode = Schema.RequiredMode.REQUIRED)
-    private Long pid;
+    private BigDecimal pid;
 
     /**
      * 路径
@@ -140,13 +141,13 @@ public class Menu extends ParentClientEntity<Menu, MenuProxy> implements ProxyEn
      * 创建人ID
      */
     @Schema(description = "创建人ID")
-    private Long createId;
+    private BigDecimal createId;
 
     /**
      * 更新人ID
      */
     @Schema(description = "更新人ID")
-    private Long updateId;
+    private BigDecimal updateId;
 
     /**
      * 是否删除
@@ -159,7 +160,7 @@ public class Menu extends ParentClientEntity<Menu, MenuProxy> implements ProxyEn
      * 租户ID
      */
     @Schema(description = "租户ID")
-    private Long tenantId;
+    private BigDecimal tenantId;
 
     /**
      *
