+++ z_doc/sql/ddl.sql
@@ -14,7 +14,7 @@ alter function update_updated_at_column() owner to postgres;
 
 create table public.z_user
 (
-    id          bigint                     not null
+    id          decimal(26,0)              not null
         primary key,
     username    varchar(32)                not null,
     pwd         varchar(512)               not null,
@@ -27,9 +27,9 @@ create table public.z_user
     email       varchar(256) default ''::character varying,
     update_time timestamp    default now() not null,
     del_flag    boolean      default false,
-    create_id   bigint       default 0,
-    update_id   bigint       default 0,
-    tenant_id   bigint       default 0
+    create_id   decimal(26,0) default 0,
+    update_id   decimal(26,0) default 0,
+    tenant_id   decimal(26,0) default 0
 );
 
 comment on table public.z_user is '系统用户表';
@@ -78,7 +78,7 @@ create index tenant_id_index
 
 create table public.z_role
 (
-    id          bigint                                     not null
+    id          decimal(26,0)                              not null
         primary key,
     role_name   varchar(64)                                not null,
     remark      varchar(255) default ''::character varying not null,
@@ -86,10 +86,10 @@ create table public.z_role
     permissions varchar(64)  default ''::character varying not null
         unique,
     update_time timestamp    default now()                 not null,
-    create_id   bigint       default 0,
-    update_id   bigint       default 0,
+    create_id   decimal(26,0) default 0,
+    update_id   decimal(26,0) default 0,
     is_lock     boolean      default false                 not null,
-    tenant_id   bigint       default 0
+    tenant_id   decimal(26,0) default 0
 );
 
 comment on table public.z_role is '系统角色表';
@@ -119,9 +119,9 @@ alter table public.z_role
 
 create table public.z_menu
 (
-    id           bigint                                     not null
+    id           decimal(26,0)                              not null
         primary key,
-    pid          bigint                                     not null,
+    pid          decimal(26,0)                              not null,
     path         varchar(255) default ''::character varying not null,
     name         varchar(64)  default ''::character varying not null,
     title        varchar(64)                                not null,
@@ -138,10 +138,10 @@ create table public.z_menu
     is_full      boolean      default false,
     is_affix     boolean      default false,
     update_time  timestamp    default now()                 not null,
-    create_id    bigint       default 0,
-    update_id    bigint       default 0,
+    create_id    decimal(26,0) default 0,
+    update_id    decimal(26,0) default 0,
     del_flag     boolean      default false                 not null,
-    tenant_id    bigint       default 0
+    tenant_id    decimal(26,0) default 0
 );
 
 comment on table public.z_menu is '系统菜单表';
@@ -195,21 +195,21 @@ alter table public.z_menu
 
 create table public.z_dept
 (
-    id              bigint                     not null
+    id              decimal(26,0)              not null
         primary key,
     name            varchar(64)                not null,
-    pid             bigint                     not null,
-    default_role_id bigint       default 0,
+    pid             decimal(26,0)              not null,
+    default_role_id decimal(26,0) default 0,
     deep            integer,
     sort            integer,
     has_children    boolean      default false,
     is_lock         boolean      default false not null,
     del_flag        boolean      default false not null,
     remark          varchar(128) default ''::character varying,
-    create_id       bigint       default 0,
-    update_id       bigint       default 0,
+    create_id       decimal(26,0) default 0,
+    update_id       decimal(26,0) default 0,
     update_time     timestamp    default now() not null,
-    tenant_id       bigint       default 0
+    tenant_id       decimal(26,0) default 0
 );
 
 comment on table public.z_dept is '部门表';
@@ -247,11 +247,11 @@ alter table public.z_dept
 
 create table public.z_user_role
 (
-    id        bigint not null
+    id        decimal(26,0) not null
         primary key,
-    user_id   bigint not null,
-    role_id   bigint not null,
-    create_id bigint default 0,
+    user_id   decimal(26,0) not null,
+    role_id   decimal(26,0) not null,
+    create_id decimal(26,0) default 0,
     constraint unique_user_role
         unique (user_id, role_id)
 );
@@ -277,11 +277,11 @@ create index role_id_index
 
 create table public.z_role_menu
 (
-    id        bigint not null
+    id        decimal(26,0) not null
         primary key,
-    role_id   bigint not null,
-    menu_id   bigint not null,
-    create_id bigint default 0,
+    role_id   decimal(26,0) not null,
+    menu_id   decimal(26,0) not null,
+    create_id decimal(26,0) default 0,
     constraint unique_role_menu
         unique (role_id, menu_id)
 );
